20251019 rename dataset class

+++ train_SIDA.py
@@ -153,7 +153,7 @@ def main(args):
     elif args.precision == "fp16":
         torch_dtype = torch.half
     model = SIDAForCausalLM.from_pretrained(
-        args.version, torch_dtype=torch_dtype, low_cpu_mem_usage=True, **model_args
+        args.version, dtype=torch_dtype, low_cpu_mem_usage=True, **model_args
     )
 
     model.config.eos_token_id = tokenizer.eos_token_id
